[Major package refactor] separate transform, analyze, failures and app
