+++ migrator/bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze/BUILD.bazel
@@ -1,0 +1,12 @@
+package(default_visibility = ["//visibility:public"])
+
+sources()
+scala_library(
+    name = "analyze",
+    srcs = [
+        ":sources",
+    ],
+    deps = [
+        "//migrator/bazel-migrator-model/src/main/java/com/wix/bazel/migrator/model:model",
+    ]
+)
