+++ migrator/bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze/CodePath.scala
@@ -1,8 +1,9 @@
-package com.wix.bazel.migrator.transform
+package com.wix.bazel.migrator.analyze
 
 import com.wix.bazel.migrator.model.SourceModule
+
 //TODO replace module and relativeSourceDirPathFromModuleRoot with SourceCodeDirPath
-private[transform] case class CodePath(module: SourceModule,
+private[migrator] case class CodePath(module: SourceModule,
                                        relativeSourceDirPathFromModuleRoot: String,
                                        filePath: String) {
   def extension: String = filePath.split('.').last
