+++ migrator/bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze/Dependency.scala
@@ -1,0 +1,3 @@
+package com.wix.bazel.migrator.analyze
+
+private [migrator] case class Dependency(codePath: CodePath, isCompileDependency: Boolean)
