--- migrator/bazel-migrator/src/main/java/com/wix/bazel/migrator/transform/Code.scala
@@ -1,7 +1,0 @@
-package com.wix.bazel.migrator.transform
-
-import com.wix.bazel.migrator.model.TestType
-
-private[transform] case class Code(codePath: CodePath, dependencies: List[Dependency] = Nil, externalSourceDependencies: Set[String] = Set.empty) {
-  def testType: TestType = TestType.from(codePath.filePath)
-}
