--- migrator/bazel-migrator/src/main/java/com/wix/bazel/migrator/transform/Dependency.scala
@@ -1,3 +1,0 @@
-package com.wix.bazel.migrator.transform
-
-private[transform] case class Dependency(codePath: CodePath, isCompileDependency: Boolean)
