+++ migrator/bazel-migrator/src/test/java/com/wix/bazel/migrator/transform/makers/BUILD.bazel
@@ -14,6 +14,7 @@ scala_library(
         "//migrator/bazel-migrator-model/src/main/java/com/wix/bazel/migrator/model",
         "//migrator/bazel-migrator/src/main/java/com/wix/bazel/migrator/transform",
         "//models/maven-model/src/main/scala/com/wixpress/build/maven",
+        "//migrator/bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze:analyze",
         "@junit_junit",
     ],
 )
