+++ migrator/bazel-migrator/src/test/java/com/wix/bazel/migrator/transform/makers/CodeMaker.scala
@@ -1,9 +1,10 @@
 package com.wix.bazel.migrator.transform.makers
 
+import com.wix.bazel.migrator.analyze
+import com.wix.bazel.migrator.analyze.{Code, Dependency}
 import com.wix.bazel.migrator.model.SourceModule
 import com.wix.bazel.migrator.model.makers.ModuleMaker.aModule
 import com.wix.bazel.migrator.transform.makers.CodePathMaker.sourceCodePath
-import com.wix.bazel.migrator.transform.{Code, Dependency}
 
 object CodeMaker {
 
@@ -12,7 +13,7 @@ object CodeMaker {
            relativeSourceDirPathFromModuleRoot: String = "src/main/java",
            dependencies: List[Dependency] = Nil,
            externalDependencies: Set[String] = Set.empty): Code =
-    Code(sourceCodePath(filePath, module, relativeSourceDirPathFromModuleRoot), dependencies, externalDependencies)
+    analyze.Code(sourceCodePath(filePath, module, relativeSourceDirPathFromModuleRoot), dependencies, externalDependencies)
 
   def testCode(filePath: String): Code = code(filePath, relativeSourceDirPathFromModuleRoot = "src/test/java", externalDependencies = Set.empty)
 }
