+++ migrator/bazel-migrator/src/test/java/com/wix/bazel/migrator/transform/makers/CodePathMaker.scala
@@ -1,12 +1,13 @@
 package com.wix.bazel.migrator.transform.makers
 
+import com.wix.bazel.migrator.analyze
+import com.wix.bazel.migrator.analyze.CodePath
 import com.wix.bazel.migrator.model.SourceModule
-import com.wix.bazel.migrator.transform.CodePath
 
 object CodePathMaker {
 
   def sourceCodePath(filePath: String,
            module: SourceModule,
            relativeSourceDirPathFromModuleRoot: String): CodePath =
-    CodePath(module, relativeSourceDirPathFromModuleRoot, filePath)
+    analyze.CodePath(module, relativeSourceDirPathFromModuleRoot, filePath)
 }
