+++ migrator/bazel-migrator/src/test/java/com/wix/bazel/migrator/transform/makers/Repo.scala
@@ -1,7 +1,7 @@
 package com.wix.bazel.migrator.transform.makers
 
+import com.wix.bazel.migrator.analyze.Code
 import com.wix.bazel.migrator.model.SourceModule
-import com.wix.bazel.migrator.transform.Code
 
 case class Repo(code: List[Code] = Nil) {
   def withCode(moreCode: Code) = Repo(moreCode +: code)
