+++ migrator/wix-bazel-migrator/BUILD.bazel
@@ -43,9 +43,8 @@ filegroup(
 
 java_binary(
     name = "migrator_cli",
-    main_class = "com.wix.bazel.migrator.MigratorApplication",
+    main_class = "com.wix.bazel.migrator.app.MigratorApplication",
     runtime_deps = [
-        "//migrator/wix-bazel-migrator/src/main/java/com/wix:agg=bazel/migrator_bazel/migrator/transform_build/maven/analysis",
-        
+        "//migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/app",
     ],
 )
