+++ migrator/wix-bazel-migrator/src/it/java/com/wix/bazel/migrator/BUILD.bazel
@@ -17,6 +17,7 @@ specs2_ite2e_test(
         "//migrator/wix-bazel-migrator:tests_dependencies",
         "//migrator/wix-bazel-migrator/src/it/java/com/wix/bazel/migrator/matchers",
         "//migrator/wix-bazel-migrator/src/main/java/com/wix:agg=bazel/migrator_bazel/migrator/transform_build/maven/analysis",
+        "//migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/app",
         "@com_github_marschall_memoryfilesystem",
         "@com_github_pathikrit_better_files_2_12",
     ],
