+++ migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze/BUILD.bazel
@@ -1,0 +1,24 @@
+package(default_visibility = ["//visibility:public"])
+
+sources()
+scala_library(
+    name = "analyze",
+    srcs = [
+        ":sources",
+    ],
+    deps = [
+        "//migrator/bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze:analyze",
+        "//dependency-resolver/maven-dependency-resolver-api/src/main/scala/com/wixpress/build/maven:maven",
+        "@com_fasterxml_jackson_core_jackson_annotations",
+        "@com_fasterxml_jackson_core_jackson_core",
+        "@com_fasterxml_jackson_core_jackson_databind",
+        "@com_fasterxml_jackson_module_jackson_module_scala_2_12",
+        "@org_slf4j_slf4j_api",
+        "@com_codota_codota_sdk_java",
+        "//migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/utils:utils",
+        "//migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/overrides:overrides",
+        "//migrator/wix-bazel-migrator/src/main/java/com/wix:agg=bazel/migrator_bazel/migrator/transform_build/maven/analysis",
+        "//repo-analyzer/maven-repo-analyzer/src/main/java/com/wix/build/zinc/analysis:analysis",
+
+    ]
+)
