+++ migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze/CompositeDependencyAnalyzer.scala
@@ -1,4 +1,4 @@
-package com.wix.bazel.migrator.transform
+package com.wix.bazel.migrator.analyze
 
 import com.wix.bazel.migrator.model.SourceModule
