+++ migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze/GeneratedCodeRegistry.scala
@@ -1,4 +1,4 @@
-package com.wix.bazel.migrator.transform
+package com.wix.bazel.migrator.analyze
 
 import com.wix.bazel.migrator.overrides.{GeneratedCodeLink, GeneratedCodeLinksOverrides}
