+++ migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze/InternalFileDepsOverridesDependencyAnalyzer.scala
@@ -1,7 +1,8 @@
-package com.wix.bazel.migrator.transform
+package com.wix.bazel.migrator.analyze
 
 import java.nio.file.Path
 
+import com.wix.bazel.migrator.analyze
 import com.wix.bazel.migrator.model.SourceModule
 import com.wix.bazel.migrator.overrides.InternalFileDepsOverridesReader
 import com.wix.build.maven.analysis.SourceModules
@@ -56,7 +57,7 @@ class InternalFileDepsOverridesDependencyAnalyzer(sourceModules: SourceModules, 
   }
 
   private def dependencyOn(isCompileDependency: Boolean)(relativeFilePath: String): Dependency =
-    Dependency(codePathFrom(relativeFilePath), isCompileDependency)
+    analyze.Dependency(codePathFrom(relativeFilePath), isCompileDependency)
 
 
 }
