+++ migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze/SourceModuleSupport.scala
@@ -1,4 +1,4 @@
-package com.wix.bazel.migrator.transform
+package com.wix.bazel.migrator.analyze
 
 import java.io.IOException
