+++ migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/app/MigratorApp.scala
@@ -1,4 +1,6 @@
-package com.wix.bazel.migrator
+package com.wix.bazel.migrator.app
+
+import com.wix.bazel.migrator.RunConfiguration
 
 trait MigratorApp extends App {
   lazy val configuration = RunConfiguration.from(args)
