+++ migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/app/MigratorApplication.scala
@@ -1,4 +1,4 @@
-package com.wix.bazel.migrator
+package com.wix.bazel.migrator.app
 
 import scala.io.Source
