+++ migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/transform/failures/BUILD.bazel
@@ -1,0 +1,13 @@
+package(default_visibility = ["//visibility:public"])
+
+sources()
+scala_library(
+    name = "failures",
+    srcs = [
+        ":sources",
+    ],
+    deps = [
+        "@com_codota_codota_sdk_java",
+        "//migrator/bazel-migrator-model/src/main/java/com/wix/bazel/migrator/model:model"
+    ]
+)
