+++ migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/utils/BUILD.bazel
@@ -11,6 +11,7 @@ scala_library(
         "//migrator/wix-bazel-migrator:main_dependencies",
         "//models/maven-model/src/main/scala/com/wixpress/build/maven",
         "@com_fasterxml_jackson_core_jackson_annotations",
+        "//migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/transform/failures:failures",
     ],
 )
