+++ migrator/wix-bazel-migrator/src/test/scala/com/wix/bazel/migrator/transform/BUILD.bazel
@@ -18,6 +18,7 @@ specs2_unit_test(
         "//migrator/bazel-migrator-model/src/main/java/com/wix/bazel/migrator/model",
         "//migrator/wix-bazel-migrator:tests_dependencies",
         "//migrator/wix-bazel-migrator/src/main/java/com/wix:agg=bazel/migrator_bazel/migrator/transform_build/maven/analysis",
+        "//migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/analyze",
         "//migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/external/registry",
         "//migrator/wix-bazel-migrator/src/main/java/com/wix/bazel/migrator/overrides",
         "//migrator/wix-bazel-migrator/src/test/scala/com/wix/bazel/migrator/external/registry",
